backend/requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
marshmallow==3.23.1
Flask-CORS==5.0.0
Flask-Limiter==3.8.0
PyJWT==2.10.1
bcrypt==4.2.1
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
uuid6==2024.7.10
python-json-logger==3.2.1
